// packages/discord/src/operations/getWebhookMessage.ts
export const GetWebhookMessageInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct(
  {
    webhook_id: Schema.String.pipe(T.PathParam()),
    webhook_token: Schema.String.pipe(T.PathParam()),
    message_id: Schema.String.pipe(T.PathParam()),
    thread_id: Schema.optional(Schema.String),
  },
).pipe(
  T.Http({
    method: "GET",
    path: "/webhooks/{webhook_id}/{webhook_token}/messages/{message_id}",
  }),
);
export type GetWebhookMessageInput = typeof GetWebhookMessageInput.Type;
// ...
export type GetWebhookMessageOutput = typeof GetWebhookMessageOutput.Type;

// The operation
export const getWebhookMessage = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: GetWebhookMessageInput,
  outputSchema: GetWebhookMessageOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));